Moved already implemented Gemini API integration to backend directory and split the same into controller and routes
