--- Package.json
@@ -1,12 +1,0 @@
-{
-  "name": "research-paper-organizer",
-  "type": "module",
-  "dependencies": {
-    "dotenv": "^16.4.0",
-    "express": "^4.19.2",
-    "node-fetch": "^3.3.2"
-  },
-  "scripts": {
-    "start": "node server.js"
-  }
-}
