+++ backend/package.json
@@ -20,6 +20,7 @@
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.17.1",
+    "node-fetch": "^3.3.2",
     "resend": "^6.0.1"
   },
   "devDependencies": {
